// app/src/lib/contracts/lb_factory/types.ts
export interface LBPair {
  token_x: string;
  token_y: string;
  bin_step: number;
  contract: {
    address: string;
    code_hash: string;
  };
}

export interface LBPairInformation {
  bin_step: number;
  lb_pair: LBPair;
  created_by_owner: boolean;
  ignored_for_routing: boolean;
}
// ...
export interface GetLBPairImplementationQuery {
  get_lb_pair_implementation: object;
}

export interface GetLBTokenImplementationQuery {
  get_lb_token_implementation: object;
}
// ...
export interface GetLBPairInformationQuery {
  get_lb_pair_information: {
    token_a: TokenType;
    token_b: TokenType;
    bin_step: number;
  };
}

export interface GetPresetQuery {
  get_preset: {
    bin_step: number;
  };
}
// ...
export interface LBPairImplementationResponse {
  lb_pair_implementation: string;
}

export interface LBTokenImplementationResponse {
  lb_token_implementation: string;
}

export interface NumberOfLBPairsResponse {
  lb_pair_number: number;
}

export interface LBPairAtIndexResponse {
  lb_pair: LBPair;
}

export interface NumberOfQuoteAssetsResponse {
  number_of_quote_assets: number;
}

export interface QuoteAssetAtIndexResponse {
  asset: string;
}

export interface IsQuoteAssetResponse {
  is_quote: boolean;
}

export interface LBPairInformationResponse {
  lb_pair_information: LBPairInformation;
}

export interface PresetResponse {
  base_factor: number;
  filter_period: number;
  decay_period: number;
  reduction_factor: number;
  variable_fee_control: number;
  protocol_share: number;
  max_volatility_accumulator: number;
  is_open: boolean;
}